+++ polaris.shopify.com/src/components/Markdown/Markdown.tsx
@@ -18,11 +18,13 @@ import {Stack} from '../../components/Stack';
 import StatusBanner from '../../components/StatusBanner';
 import TipBanner from '../../components/TipBanner';
 import {Lede} from '../../components/Lede';
+import Subnav from '../../components/Subnav';
 import {SideBySide} from './components/SideBySide';
 import YoutubeVideo from '../YoutubeVideo';
 import {DoDont, Do, Dont} from './components/DoDont';
 import {Heading} from '../../components/Heading';
 import PatternsExample from '../../components/PatternsExample';
+import {Card} from '../../components/Card';
 import RichCardGrid from '../RichCardGrid';
 import Tooltip from '../Tooltip';
 import Icon from '../Icon';
@@ -30,10 +32,12 @@ import {FeaturedCardGrid} from '../FeaturedCardGrid';
 import {useCopyToClipboard} from '../../utils/hooks';
 import {Colors} from './components/Colors';
 
-const CodeVisibilityContext = createContext<[boolean, (arg: boolean) => void]>([
-  false,
-  () => {},
-]);
+const CodeVisibilityContext = createContext<
+  [
+    boolean | undefined,
+    React.Dispatch<React.SetStateAction<boolean>> | undefined,
+  ]
+>([undefined, undefined]);
 
 const useCodeVisibility = () => useContext(CodeVisibilityContext);
 
@@ -46,16 +50,17 @@ export const CodeVisibilityProvider: ComponentType<
     /**
      * This value will be overwritten by the outer-most context provider
      */
-    setShowCode: (arg: boolean) => void;
+    setShowCode: React.Dispatch<React.SetStateAction<boolean>>;
   }>
 > = ({children, showCode, setShowCode}) => {
   // Attempt to inherit from parent context if it's set
   const [codeVisibleFromContext, setShowCodeFromContext] = useCodeVisibility();
-
   return (
     <CodeVisibilityContext.Provider
       value={[
-        codeVisibleFromContext ?? showCode,
+        typeof codeVisibleFromContext !== 'undefined'
+          ? codeVisibleFromContext
+          : showCode,
         setShowCodeFromContext ?? setShowCode,
       ]}
     >
@@ -84,7 +89,6 @@ const FencedCodeWithVisibilityToggle: ComponentType<
   meta: {title, type, previewContext, sandboxContext, isActionsVisible} = {},
 }) => {
   const [showCode, setShowCode] = useCodeVisibility();
-
   if (type === 'livePreview') {
     return (
       <PatternsExample
@@ -95,7 +99,9 @@ const FencedCodeWithVisibilityToggle: ComponentType<
         }}
         isCodeVisible={showCode}
         isActionsVisible={isActionsVisible}
-        onCodeVisibilityToggle={() => setShowCode(!showCode)}
+        onCodeVisibilityToggle={() => {
+          setShowCode?.(!showCode);
+        }}
         title={title ?? 'Pattern Name'}
       />
     );
@@ -221,6 +227,7 @@ function Markdown<
         InlineGrid,
         SideBySide,
         Grid,
+        Card,
         FeaturedCardGrid,
         YoutubeVideo,
         DoDont,
@@ -229,6 +236,7 @@ function Markdown<
         StatusBanner,
         TipBanner,
         Lede,
+        Subnav,
         RichCardGrid,
         Colors,
         Tip: ({children}) => (
@@ -262,7 +270,7 @@ function Markdown<
 }
 
 export const HeadingWithCopyButton = forwardRef(
-  ({id, children, ...props}, ref) => {
+  ({id, children, className, ...props}, ref) => {
     const origin =
       typeof window !== 'undefined'
         ? window.location.origin
@@ -291,7 +299,11 @@ export const HeadingWithCopyButton = forwardRef(
         as="h1"
         id={id}
         {...props}
-        className={[styles.MarkdownHeading, styles[`Heading-${props.as}`]]}
+        className={[
+          styles.MarkdownHeading,
+          styles[`Heading-${props.as}`],
+          className,
+        ]}
         ref={ref}
       >
         {children}