+++ polaris.shopify.com/src/utils/various.ts
@@ -1,5 +1,12 @@
 import siteJson from '../../.cache/site.json';
 import {PatternFrontMatter, SiteJSON} from '../types';
+import type {BreakpointsAlias} from '@shopify/polaris-tokens';
+import {breakpointsAliases} from '@shopify/polaris-tokens';
+
+export function isObject(value: any) {
+  const type = typeof value;
+  return value != null && (type === 'object' || type === 'function');
+}
 
 interface PatternJSON {
   [key: string]: {
@@ -55,6 +62,10 @@ export const className = (...classNames: ClassName[]): string => {
     .join(' ');
 };
 
+export const variationName = (name: string, value: string) => {
+  return `${name}${value.charAt(0).toUpperCase()}${value.slice(1)}`;
+};
+
 export const toPascalCase = (str: string): string =>
   (str.match(/[a-zA-Z0-9]+/g) || [])
     .map((w) => `${w.charAt(0).toUpperCase()}${w.slice(1)}`)
@@ -69,3 +80,61 @@ export const deslugify = (str: string): string =>
 export const capitalize = (str: string) => {
   return str.charAt(0).toUpperCase() + str.slice(1);
 };
+
+/**
+ * Given params like so:
+ * (
+ *   'button',
+ *   'padding',
+ *   'spacing',
+ *   {
+ *     sm: "4",
+ *     lg: "6"
+ *   }
+ * )
+ * Converts it to an object like so:
+ * {
+ *   '--pc-button-padding-sm': 'var(--p-spacing-4)',
+ *   '--pc-button-padding-lg': 'var(--p-spacing-6)'
+ * }
+ *
+ */
+type ResponsivePropConfig<T = string> = {
+  [Breakpoint in BreakpointsAlias]?: T;
+};
+export type ResponsiveProp<T> = T | ResponsivePropConfig<T>;
+type ResponsiveVariables<T> = {
+  [Breakpoint in `${string}-${BreakpointsAlias}`]?: T;
+};
+
+export function getResponsiveProps<T = string>(
+  componentName: string,
+  componentProp: string,
+  tokenSubgroup: string,
+  responsiveProp?: ResponsiveProp<T>,
+): ResponsiveVariables<T> {
+  if (!responsiveProp) return {};
+
+  let result: ResponsivePropConfig;
+
+  if (!isObject(responsiveProp)) {
+    result = {
+      [breakpointsAliases[0]]: `var(--p-${tokenSubgroup}-${responsiveProp})`,
+    };
+  } else {
+    result = Object.fromEntries(
+      Object.entries(responsiveProp).map(([breakpointAlias, aliasOrScale]) => [
+        breakpointAlias,
+        `var(--p-${tokenSubgroup}-${aliasOrScale})`,
+      ]),
+    );
+  }
+
+  // Prefix each responsive key with the correct token name
+  return Object.fromEntries(
+    Object.entries(result).map(([breakpointAlias, value]) => [
+      `--pc-${componentName}-${componentProp}-${breakpointAlias}`,
+      value,
+    ]),
+  ) as unknown as ResponsiveVariables<T>;
+}